Refactor map generating
